Support node back to v0.12
