+++ package.json
@@ -18,6 +18,8 @@
   "dependencies": {
     "Base64": "0.3.0",
     "babel-eslint": "5.0.0",
+    "babel-preset-es2015": "6.6.0",
+    "babel-register": "6.7.2",
     "broccoli": "0.16.9",
     "broccoli-asset-rev": "2.4.1",
     "broccoli-babel-transpiler": "5.5.0",
@@ -25,24 +27,27 @@
     "broccoli-concat": "2.2.0",
     "broccoli-env": "0.0.1",
     "broccoli-funnel": "1.0.1",
-    "broccoli-lint-eslint": "1.2.0",
+    "broccoli-lint-eslint": "2.2.0",
     "broccoli-merge-trees": "1.1.1",
+    "broccoli-plugin": "1.2.1",
     "broccoli-sane-watcher": "1.1.4",
     "broccoli-static-compiler": "0.2.2",
+    "broccoli-uglify-js": "0.1.3",
     "broccoli-uglify-sourcemap": "1.1.1",
     "escope": "3.3.0",
     "eslint-config-nightmare-mode": "2.3.0",
     "estraverse-fb": "1.3.1",
+    "fake-xml-http-request": "1.4.0",
     "js-string-escape": "1.0.0",
     "loader.js": "3.5.0",
     "node-fetch": "1.3.3",
     "pretender": "0.10.1",
     "qunitjs": "1.20.0",
+    "route-recognizer": "0.1.9",
     "rsvp": "3.1.0",
     "testem": "0.9.11",
     "whatwg-fetch": "0.10.1",
-    "yuidocjs": "0.10.0",
-    "broccoli-uglify-js": "0.1.3"
+    "yuidocjs": "0.10.0"
   },
   "devDependencies": {
     "aws-sdk": "2.2.44",
